+++ CHANGELOG.md
@@ -69,6 +69,7 @@ This option is new for VectorCAST 23 sp2
 ## [1.0.8] - 2023-10-06
 
 - Added new right click menu choices for .vce (open environment) and .env (build environment) files #30
+- Added new right click menu choice for test explorer tree to allow you to insert Basis Path and ATG tests #31
 - Improved the open settings feature to filter properly and default to the 'Workspace' tab
 
 ### Bug Fixes